gcp/gcf/pub-test/requirements.txt
google-cloud-pubsub >= 2.2.0
google-cloud-bigquery >= 2.7.0
google-cloud-storage >= 1.35.0